UNLOST/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Login
Flask-Bcrypt
Authlib
requests
pyotp
Flask-Mail
